requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.2
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
